Reviewed Sonar issues
- Disabled generics wildcard checks (java:S1452)
- Fixed ZoneOffset seconds generation (java:S2184)
- Improved `ArrayGenerator` default values test (java:S5961)

+++ src/main/java/dev/orne/test/rnd/generators/ZoneOffsetGenerator.java
@@ -60,7 +60,7 @@ extends AbstractTypedGenerator<ZoneOffset> {
     public @NotNull ZoneOffset randomValue() {
         final long seconds = RandomUtils.nextLong(
                 0,
-                ZoneOffset.MAX.getTotalSeconds() - ZoneOffset.MIN.getTotalSeconds())
+                (long) ZoneOffset.MAX.getTotalSeconds() - (long) ZoneOffset.MIN.getTotalSeconds())
                 + ZoneOffset.MIN.getTotalSeconds();
         return ZoneOffset.ofTotalSeconds((int) seconds);
     }
